services/catalog-service/requirements.txt
django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.11
django-filter>=23.5
pillow==10.2.0
drf-spectacular==0.27.1
gunicorn==21.2.0
uvicorn==0.27.0